Add: qvarn.Stopwatch class
The Python cProfile profile doesn't seem to catch callbacks from
gunicorn, so we add a Stopwatch class, which is a context manager and
logs how long the context waa. This ia poor substitute for a good
profiler, but better than nothing.

Also add use of the stopwatch in a number of places of interest.