Change: lib.py: globals vars in V, manage list of qvarns, tokens
This commit is preparation for large changes to Qvarn and yarns, for
the fine-grained access control implementation. The changes are
necessary because the old code was so much of a mess.

vars is a Python built-in so we should not use that name. V is a
better name: it doesn't conflict, and it's short, and visually
distinctive. It is not indicative of a vendetta.

Change the way we store information about Qvarn instances we start, in
V. This makes it easier to manage multiple Qvarns. Ditto for Qvarn and
user specific tokens.

Add helper functions to lib.py for making requests to a specific Qvarn
and storing the results in V. Most importantly, the new request
function does most of the work, simplifyin the IMPLEMENTS sections
something wonderful.