Change: listen on all interfaces
Previously, haproxy and Qvarn were running on the same host. Now we
will be deploying a separate haproxy node, and that two nodes will
need to communicate across the network. Thus, we make Qvarn be started
so that it listens on all interfaces, instead of 127.0.0.1.

Clearly, this means the connections between haproxy and the Qvarn
nodes need to be secured in some way.